Release 5.3.4

+++ views/js/pciCreator/dev/audioRecordingInteraction/runtime/js/providers/webAudio.js
@@ -169,17 +169,14 @@ define([
             },
 
             /**
-             * Close the audio context and destroy created assets
+             * Destroy created assets
              */
             destroy: function destroy() {
-                recorderWorker.terminate();
-                recorderWorker = null;
-
-                if (getAudioContext()) {
-                    return getAudioContext().close().then(function() {
-                        audioNodes = {};
-                    });
+                if (recorderWorker) {
+                    recorderWorker.terminate();
+                    recorderWorker = null;
                 }
+                audioNodes = {};
             }
         };
